EVEREST-980 | (Bug fix) Ensure install approves the correct InstallPlan (#202)
install command uses the InstallPlan provided in the InstallPlanRef of the subscription. However, OLM may change this over time if a new InstallPlan is available. This means that running install twice may lead to the installation of an undesired version and break Everest installation.

This PR fixes this by checking the `installedCSV` of the Subscription and re-using that. If empty, it will use the `startingCSV` or the InstallPlanRef.