EVEREST-107 remove index.html
We don't want to version control this file because it has no meaning and
will always be overwritten by the build process. It only existed so that one
doesn't have to build the UI to run the API server. By creating it in the
CI workflows we can run the BE CI pipelines without needing to have it
in git.