EVEREST-1182 fix install of everest operator
We introduced this phase check because OLM has a bug where it sometimes
creates duplicate InstallPlans for the same CSV and we found a few cases
where the duplicate InstallPlan wasn't reconciled correctly and
abandoned by OLM. This abandoned InstallPlan was missing the status
field meaning it was also missing the necessary plan to install the
operator. Approving this InstallPlan would cause the operator to never
be installed. By checking the phase we make sure we will be approving an
InstallPlan that is actually ready to be approved. See
https://github.com/operator-framework/kubectl-operator/issues/13 for
more details on a similar issue.