EVEREST-1403 | Fix broken uninstallation after 1.2.0 API migration (#632)
As a part of the 1.2.0 upgrade, we perform a migration of BackupStorages and MonitoringConfigs.
Prior to 1.2.0, the BackupStorages have a `percona.com/cleanup-secrets` finalizer that was intended to remove Secrets in the DB namespaces.

Uninstallation gets stuck because this finalizer is no longer handled by the operator

**Solution:**
* Remove the finalizer during the migration, since the Secrets in the DB namespaces are now owned by the newly created local BackupStorage
* Also fix the uninstall code to list BackupStorages/MonitoringConfigs from DB namespaces.

Signed-off-by: Mayank Shah <mayank.shah@percona.com>
(cherry picked from commit 4b2d716ccf098414c0a577cd8813525adabcb847)