EVEREST-1403 | Run 1.2.0 API migration as a part of the post-upgrade step (#633)
In #632 we added some code to remove a finalizer from BackupStorages in everest-system namespace to unblock uninstallation. However, that code is run before the operator has upgraded. As a result, the operator adds back this finalizer.

To fix this problem, we run the migration after the Everest operator upgrade step, i.e, as a part of the post-upgrade tasks.

Signed-off-by: Mayank Shah <mayank.shah@percona.com>
(cherry picked from commit 86d127893e04d1ce08cb416172b9525733be6383)