EVEREST-107 fix cliend-go dep
https://github.com/percona/everest-operator/pull/581 introduced a
dependency change. The victoria metrics module imported in that PR
changed the k8s.io/client-go dependency from v0.31.1 to
v12.0.0+incompatible. In order to fix that we pinned the dependencies
with a replace block.

Since this repo imports the everest-operator module, we need to do
the same otherwise we can't build.