EVEREST-1752 | add support for more helm upgrade CLI flags (#937)
There’s an issue with Helm where during upgrades, if you provide values using --set (or equivalent flags), it will reset the values back to the defaults built into the chart.

To tweak this behaviour, helm offers the following 3 flags: reuse-values, reset-values and reset-then-reuse-values.

Each of these options changes the way values are applied during upgrades. Since we provide an option to change the Helm values from the CLI, we need to provide these flags as well.

Signed-off-by: Mayank Shah <mayank.shah@percona.com>