EVEREST-107 use ROBOT_TOKEN instead of GITHUB_TOKEN
When you use the repository's GITHUB_TOKEN to perform tasks, events
triggered by the GITHUB_TOKEN, with the exception of workflow_dispatch
and repository_dispatch, will not create a new workflow run. This
prevents you from accidentally creating recursive workflow runs. For
example, if a workflow run pushes code using the repository's
GITHUB_TOKEN, a new workflow will not run even when the repository
contains a workflow configured to run when push events occur.
Source: https://docs.github.com/en/actions/security-for-github-actions/security-guides/automatic-token-authentication#using-the-github_token-in-a-workflow

This means that the updates to the PRs wouldn't trigger the other
workflows that usually run in the PRs thus merging them was blocked
because the checks didn't pass. Changing the token to the ROBOT_TOKEN
removes this limitation.