EVEREST-107 fix auto-update-base when PR is failing checks
The `mergeable_state` is an undocumented (not production-ready) field so
we shouldn't use it. Also, it didn't allow us to understand if a PR in
the 'blocked' state (PR checks running or failing) needed an update or
not. By using the `mergeable` field, we can detect merge conflicts and
then compare the base branch hash reported by the PR object with the
actual base branch hash and trigger the update accordingly.